Move `ansible_hosts` to example_06
After reviewing examples 04, 05, and 06, I am moving the creation of the
ansible_hosts file to `example_06`. Initially, we asked the user for the
details and recreated the file each time. However, not only was that
tedious for the user, it also took focus away from the purpose of
ansible.

Example 06 is where the file will be used without a helper script to run
the commands for you (e.g., where the training wheels come off). It is
here that it makes the most sense to put the file (of the three
examples).

It was very tempting to place the file in a default location like
/etc/ansible/hosts or $HOME. However, I hestiate storying anything
outside of this repository (so all data can be cleanly removed if the
user wishes). The user can move this file to a default location if
she/he wishes.