Add third role_example (role: app_server_django)
Because there won't be time for the additional components, no effort has gone
in to automate the django process. Simply manually running the Django runserver
and refreshing the page after this deploy is finished should be sufficient.

Extra work (such as installing gunicorn and also another role for 'supervisor')
was cut from the presentation because there simply won't be time to review
everything. This is the last of the demos (currently). Future versions will probably
add a gunicorn, supervisor, and 'small deployment process' roles.