Fix gunicorn dependency
The django package that I uploaded was already configured and tested with
gunicorn before I canned it. With over 20 examples already (even if each took
just 3 minutes), we are out of time.  Instead of having to back out a
compressed file, I am just adding gunicorn until after the presentation.