Fix symlinks for playbook_examples
When checking out a fresh repo (to more ruggedly test the repo), I discovered
that the two symlinks to the ansible.cfg and ansible_hosts were not working.
This means that none of the playbook_examples would have worked unless someone
knew to make these links (I'm trying to make that easy so one doesn't have to
think about such things until the core concepts are covered). Fixing now.