Add -vvv explanation to README.txt
This is a partial "change of mind" from a previous pull request covnersation
with https://github.com/msabramo. In that conversation he suggested adding
the '-vvv' options to the command. I wanted to keep this output short and
easier to digest.

But, I realize that there is a difference between this example and the previous
one. The previous example was using a module '-m' and this example sending a command '-a'.

So, verbose should be explained again to show that verbose can be used in
either scneario. In future commits these two examples order will be switched
(because the command makes more sense to introduce first before the module.)